+++ pkg/controller/infrastructure/actuator_migrate.go
@@ -1,0 +1,21 @@
+// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package infrastructure
+
+import (
+	"context"
+
+	extensioncontroller "github.com/gardener/gardener/extensions/pkg/controller"
+	"github.com/gardener/gardener/extensions/pkg/util"
+	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	"github.com/go-logr/logr"
+
+	"github.com/gardener/gardener-extension-provider-alicloud/pkg/apis/alicloud/helper"
+)
+
+// Migrate implements infrastructure.Actuator.
+func (a *actuator) Migrate(ctx context.Context, log logr.Logger, infra *extensionsv1alpha1.Infrastructure, _ *extensioncontroller.Cluster) error {
+	return util.DetermineError(a.cleanupTerraformerResources(ctx, log, infra), helper.KnownCodes)
+}
